Extract the error_get_last() handling into a separate Exception class.
In all cases (Horde_Feed, Horde_Xml_Element, Horde_Yaml, and Horde_Http) where
we throw the result from error_get_last() into the exception class we know
exactly that we are doing so.

Rather than having the Exception class trying to determine what kind of
parameters it might have received we can use a separate class for that.